💚 chore(ci): remove comment creation step from helm-e2e-test workflow
Removed the step responsible for creating comments with test results in pull requests from the GitHub Actions workflow for helm end-to-end tests. This change simplifies the workflow by eliminating the use of the `peter-evans/create-or-update-comment` action, which was conditionally executed based on the pull request and test outcome. The workflow now focuses solely on the execution and artifact uploading of Cypress end-to-end tests, potentially reducing unnecessary comments and clutter in pull requests.