📝 docs: update README with Ukrainian support badge
Add a badge to the README indicating that the project is made by Ukrainian. Additionally, include a "Stand with Ukraine" banner at the end of the file to show solidarity and support. This enhancement aims to acknowledge and promote the contributors' background while also creating awareness regarding the ongoing situation in Ukraine.